requirements.txt
anthropic>=0.18.0
langgraph>=0.0.20
python-dotenv>=1.0.0